### 2018-10-19 #1033 @bastistician
Scanning through https://bit.ly/joss-reviewers, I found @nuest and @wrathematics to be potentially suitable reviewers for the R package **polyCub**.
