### 2018-10-31 #1033 @wrathematics
Sorry for the delay. I keep messing up my email filters for JOSS, so I haven't been seeing the pings...

I believe I can review this package.
