### 2018-11-01 #1056 @whedon
[REVIEW]: polyCub: An R package for Integration over Polygons
**Submitting author:** @bastistician (<a href="http://orcid.org/0000-0002-1791-9449">Sebastian Meyer</a>)
**Repository:** <a href="https://github.com/bastistician/polyCub" target ="_blank">https://github.com/bastistician/polyCub</a>
**Version:** 0.7.0
