### 2018-11-14 #1056 @wrathematics
Sorry this took me so long. I think everything basically looks good. A few very minor comments:

* The README.md file describes how to contribute, report problems, and seek support. But I think it's somewhat popular to have a CONTRIBUTING.md file. Personally I prefer having that information in a README, but you might consider creating one.
* I found one minor compiler warning. I submitted an issue on your repo, but as I say, it's quite minor so you may close the issue if you wish to ignore.
* Your README.md is quite thorough. You might consider turning it into a package vignette.
* I would personally like to see some benchmarks (perhaps in a future vignette). Even if you just compared the performance/accuracy tradeoffs that you outline in the paper, I think it would be very helpful.

As far as the JOSS review is concerned, I don't consider any of these issues blockers.
