### 2018-11-14 #1056 @karthik
Thanks very much @wrathematics!
@bastistician Please let me know if you want to address these optional fixes before acceptance.
