### 2018-11-15 #1056 @bastistician
Thanks a lot for the review, @wrathematics! (BTW: I wouldn't judge two weeks as a long review period.)
I really like your suggestion of turning the README into a package vignette as well as including some benchmarks therein. Let me do that before acceptance (@karthik).
