### 2018-11-23 d96a273
cubature now integrates the Cuba libary
