+++ DESCRIPTION
@@ -48,4 +48,4 @@ Suggests:
     rgeos,
     gpclib,
     cubature
-RoxygenNote: 6.1.0
+RoxygenNote: 6.1.1
