### 2019-01-22 ea0f58c
use types array in R_CMethodDef
Fixes #1
