+++ src/init.c
@@ -1,7 +1,7 @@
 /*******************************************************************************
  * Registering native routines (entry points in compiled code)
  *
- * Copyright (C) 2017 Sebastian Meyer
+ * Copyright (C) 2017,2019 Sebastian Meyer
  *
  * This file is part of the R package "polyCub",
  * free software under the terms of the GNU General Public License, version 2,
@@ -9,14 +9,22 @@
  ******************************************************************************/
 
 #include <stdlib.h> // for NULL
+#include <Rinternals.h> // for SEXP types
 #include <R_ext/Rdynload.h>
 
 #include "polyCub.SV.h"
 #include "polyCub.iso.h"
 
+// types array (could be omitted)
+static R_NativePrimitiveArgType C_polygauss_t[] = {
+    REALSXP, REALSXP, REALSXP, REALSXP, REALSXP, REALSXP, REALSXP,
+    /*L, M, N:*/ INTSXP, INTSXP, INTSXP,
+    /*results:*/ REALSXP, REALSXP, REALSXP
+};
+
 static const R_CMethodDef CEntries[] = {
-    {"C_polygauss", (DL_FUNC) &C_polygauss, 13},
-    {NULL, NULL, 0}
+    {"C_polygauss", (DL_FUNC) &C_polygauss, 13, C_polygauss_t},
+    {NULL, NULL, 0, NULL}
 };
 
 void R_init_polyCub(DllInfo *dll)
