### 2019-01-24 adf405f
add required vignette builders
