+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Package: polyCub
 Title: Cubature over Polygonal Domains
 Version: 0.7.0.9000
-Date: 2018-11-23
+Date: 2018-01-24
 Authors@R: c(
     person("Sebastian", "Meyer",
            email = "seb.meyer@fau.de",
@@ -47,5 +47,8 @@ Suggests:
     statmod,
     rgeos,
     gpclib,
-    cubature
+    cubature,
+    knitr,
+    rmarkdown
 RoxygenNote: 6.1.1
+VignetteBuilder: knitr, rmarkdown
