### 2019-01-24 536c195
create package logo in standard location (man/figures) and size
