### 2019-01-24 fd9441c
make vignette, drop logo and README content, restructure
