+++ vignettes/polyCub.Rmd
@@ -1,41 +1,18 @@
-```{r setup, include = FALSE}
-## Storing figures in man/figures would inflate the package.
-## Instead we .Rbuildignore the generated figures and use the full GitHub path
-## when referencing the figures in the resulting markdown file
-knitr::opts_knit$set(
-  base.url = "https://raw.githubusercontent.com/bastistician/polyCub/master/"
-)
-knitr::opts_chunk$set(
-  collapse = TRUE,
-  comment = "#>",
-  fig.path = "figures/", dev = "png", dpi = 96,
-  fig.width = 4.5, fig.height = 3.75, dev.args = list(pointsize = 10)
-)
-```
-```{R example-f, include = FALSE}
-f <- function (s, sigma = 5)
-{
-    exp(-rowSums(s^2)/2/sigma^2) / (2*pi*sigma^2)
-}
-```
-```{R example-polygon, include = FALSE}
-hexagon <- list(
-    list(x = c(7.33, 7.33, 3, -1.33, -1.33, 3),
-         y = c(-0.5, 4.5, 7, 4.5, -0.5, -3))
-)
-```
-```{R logo, fig.width = 1.3, fig.height = 1.5, include = FALSE}
-.f_inside_hexagon <- function (s, ...) {
-    inside <- sp::point.in.polygon(s[,1], s[,2], hexagon[[1]]$x, hexagon[[1]]$y)
-    ifelse(inside, f(s, ...), NA_real_)
-}
-par(bg = "transparent", ann = FALSE, bty = "n", xaxt = "n", yaxt = "n", mar = c(0,0,0,0))
-polyCub::plotpolyf(hexagon, .f_inside_hexagon, use.lattice = FALSE,
-                   xlim = range(hexagon[[1]]$x), ylim = range(hexagon[[1]]$y))
-text(3, 2, labels = "\u222B", cex = 5)  # Unicode integral symbol
+---
+title: "Getting started with polyCub"
+author: "Sebastian Meyer"
+date: "`r Sys.Date()`"
+output: rmarkdown::html_vignette
+vignette: >
+  %\VignetteIndexEntry{Getting started with polyCub}
+  %\VignetteEngine{knitr::rmarkdown}
+---
+
+```{R setup, include = FALSE, purl = FALSE}
+knitr::opts_chunk$set(collapse = TRUE, comment = "#>")
 ```
 
-# polyCub <img src="`r paste0(knitr::opts_knit$get("base.url"), "figures/logo-1.png")`" align="right" />
+<img src="../man/figures/logo.png" align="right" alt="" width="120" />
 
 The R package **polyCub** implements *cubature* (numerical integration)
 over *polygonal* domains.
@@ -43,29 +20,10 @@ It solves the problem of integrating a continuously differentiable
 function f(x,y) over simple closed polygons.
 
 For the special case where the domain is rectangular with sides parallel
-to the axes (such as a bounding box), the packages
+to the axes (such as a bounding box), the package
 [**cubature**](https://CRAN.R-project.org/package=cubature)
-and [**R2Cuba**](https://CRAN.R-project.org/package=R2Cuba)
-are more appropriate (cf.
+is more appropriate (cf.
 [`CRAN Task View: Numerical Mathematics`](https://CRAN.R-project.org/view=NumericalMathematics)).
-
-
-## Installation
-
-You can install
-[polyCub from CRAN](https://CRAN.R-project.org/package=polyCub)
-via:
-
-```R
-install.packages("polyCub")
-```
-
-To install the development version from the GitHub repository, use:
-
-```R
-## install.packages("remotes")
-remotes::install_github("bastistician/polyCub")
-```
 
 
 ## Motivation
[…]
