### 2019-01-29 a56275e
revise and simplify
