+++ vignettes/polyCub.Rmd
@@ -19,77 +19,63 @@ over *polygonal* domains.
 It solves the problem of integrating a continuously differentiable
 function f(x,y) over simple closed polygons.
 
-For the special case where the domain is rectangular with sides parallel
-to the axes (such as a bounding box), the package
+For the special case of a rectangular domain along the axes, the package
 [**cubature**](https://CRAN.R-project.org/package=cubature)
 is more appropriate (cf.
 [`CRAN Task View: Numerical Mathematics`](https://CRAN.R-project.org/view=NumericalMathematics)).
 
 
-## Motivation
-
-The **polyCub** package evolved from the need to evaluate integrals of
-so-called spatial interaction functions (e.g., a Gaussian or power-law
-kernel) over the observation region of a spatio-temporal point process
-(Meyer et al, 2012, *Biometrics*,
-<https://doi.org/10.1111/j.1541-0420.2011.01684.x>).
-Such an observation region is described by a polygonal boundary,
-representing, for example, the shape of a country or administrative
-district.
-
-
 ## Polygon representations
+
+The integration domain is described by a polygonal boundary
+(or multiple polygons, including holes).
+Various R packages for spatial data analysis provide classes for polygons.
+The implementations differ in vertex order (which direction represents a hole)
+and if the first vertex is repeated.
 
 All of **polyCub**'s cubature methods understand
 
-* `"owin"` from **spatstat**,
+* `"owin"` from package [**spatstat**](https://CRAN.R-project.org/package=spatstat),
 
-* `"gpc.poly"` from package **rgeos** (or **gpclib**), and
+* `"gpc.poly"` from [**rgeos**](https://CRAN.R-project.org/package=rgeos)
+  (or [**gpclib**](https://CRAN.R-project.org/package=gpclib)), and
 
-* `"SpatialPolygons"` from package **sp**.
+* `"SpatialPolygons"` from package [**sp**](https://CRAN.R-project.org/package=sp).
 
-Some also support a plain list of lists of vertex coordinates (`"xylist"`).
-
-Note that the integration domain may consist of more than one polygon
-(including holes).
+Internally, **polyCub** uses its auxiliary `xylist()` function to extract
+a plain list of lists of vertex coordinates from these classes,
+such that vertices are ordered anticlockwise and the first vertex is not
+repeated (i.e., the `"owin"` convention).
 
 
 ## Cubature methods
 
 The following cubature methods are implemented in **polyCub**:
 
-1. **Product Gauss cubature** as proposed by Sommariva and Vianello
-   (2007, *BIT Numerical Mathematics*,
-   <https://doi.org/10.1007/s10543-007-0131-2>).
+1. `polyCub.SV()`: Product Gauss cubature
 
-2. The simple **two-dimensional midpoint rule** via `as.im.function()` from
-   the [**spatstat**](https://CRAN.R-project.org/package=spatstat) package.
+2. `polyCub.midpoint()`: Two-dimensional midpoint rule
   
-3. For **radially symmetric functions** $f(x,y) = f_r(\lVert(x-x_0,y-y_0)\rVert)$,
-   numerical integration can be made much more efficient via line
-   `integrate()` along the boundary of the polygonal domain
-   (Meyer and Held, 2014, *The Annals of Applied Statistics*,
-   <https://doi.org/10.1214/14-AOAS743>, Supplement B, Section 2.4).
+3. `polyCub.iso()`: Efficient integration of radially symmetric functions
+   $f(x,y) = f_r(\lVert(x-x_0,y-y_0)\rVert)$
 
-4. For **bivariate Gaussian densities**, integrals over polygons can be
-   solved accurately (but slowly) based on a triangulation of the domain
-   (via `tristrip()` from the
[…]
