### 2019-01-29 8d77732
remove details (refer to new vignette)
