+++ README.md
@@ -1,20 +1,14 @@
 
-
-
-
-
-# polyCub <img src="https://raw.githubusercontent.com/bastistician/polyCub/master/figures/logo-1.png" align="right" />
+# polyCub <img src="man/figures/logo.png" align="right" alt="" width="120" />
 
 The R package **polyCub** implements *cubature* (numerical integration)
 over *polygonal* domains.
 It solves the problem of integrating a continuously differentiable
 function f(x,y) over simple closed polygons.
 
-For the special case where the domain is rectangular with sides parallel
-to the axes (such as a bounding box), the packages
+For the special case of a rectangular domain along the axes, the
 [**cubature**](https://CRAN.R-project.org/package=cubature)
-and [**R2Cuba**](https://CRAN.R-project.org/package=R2Cuba)
-are more appropriate (cf.
+package is more appropriate (cf.
 [`CRAN Task View: Numerical Mathematics`](https://CRAN.R-project.org/view=NumericalMathematics)).
 
 
@@ -36,297 +30,52 @@ remotes::install_github("bastistician/polyCub")
 ```
 
 
-## Motivation
+## Cubature methods
 
-The **polyCub** package evolved from the need to evaluate integrals of
-so-called spatial interaction functions (e.g., a Gaussian or power-law
-kernel) over the observation region of a spatio-temporal point process
-(Meyer et al, 2012, *Biometrics*,
-<https://doi.org/10.1111/j.1541-0420.2011.01684.x>).
-Such an observation region is described by a polygonal boundary,
-representing, for example, the shape of a country or administrative
-district.
+1. General-purpose **product Gauss cubature** (Sommariva and Vianello, 2007,
+   *BIT Numerical Mathematics*, <https://doi.org/10.1007/s10543-007-0131-2>)
 
-The integration task could be simplified by either assuming a trivial
-kernel, such as f(x,y)=1, or by simply replacing the polygonal with a
-rectangular domain, such as the bounding box of the polygon.
-However, these crude approximations can be avoided by using efficient
-numerical integration methods for polygonal domains:
-
-1. *Product Gauss cubature* as proposed by Sommariva and Vianello
-   (2007, *BIT Numerical Mathematics*,
-   <https://doi.org/10.1007/s10543-007-0131-2>).
-
-2. The simple *two-dimensional midpoint rule* via `as.im.function()` from
-   the [**spatstat**](https://CRAN.R-project.org/package=spatstat) package.
+2. Simple **two-dimensional midpoint rule**
+   (via [**spatstat**](https://CRAN.R-project.org/package=spatstat))
   
-3. For *radially symmetric functions* f(x,y) = f_r(||(x-x_0,y-y_0)||),
-   numerical integration can be made much more efficient via line
-   `integrate()` along the boundary of the polygonal domain
+3. **Adaptive cubature for radially symmetric functions**
+   f(x,y) = f_r(||(x-x_0,y-y_0)||)
+   via line `integrate()` along the polygon boundary
    (Meyer and Held, 2014, *The Annals of Applied Statistics*,
-   <https://doi.org/10.1214/14-AOAS743>, Supplement B, Section 2.4).
+   <https://doi.org/10.1214/14-AOAS743>, Supplement B, Section 2.4)
 
-4. For *bivariate Gaussian densities*, integrals over polygons can be
-   solved accurately (but slowly) based on a triangulation of the domain
-   (via `tristrip()` from the
-   [**gpclib**](https://CRAN.R-project.org/package=gpclib) package)
[…]
