### 2019-01-29 1859057
add runtime comparison (skipped on CRAN)
