### 2019-01-29 d4fb98b
add microbenchmark (used in vignette)
