### 2019-01-30 a864b33
add NEWS entry
