### 2019-01-30 7523a53
simplify structure (basic usage first)
