### 2019-01-30 047c327
update and cleanup
