+++ paper/paper.bib
@@ -83,21 +83,10 @@
 
 @Manual{R:cubature,
   title  = {\texttt{cubature}: Adaptive Multivariate Integration over Hypercubes},
-  author = {Balasubramanian Narasimhan and Steven G. Johnson},
+  author = {Balasubramanian Narasimhan and Steven G. Johnson and Thomas Hahn and Annie Bouvier and Kiên Kiêu},
   year   = {2018},
-  note   = {R package version 1.4},
+  note   = {R package version 2.0.3},
   url    = {https://CRAN.R-project.org/package=cubature},
-}
-
-@Article{pebesma2018,
-  author  = {Edzer Pebesma},
-  title   = {Simple Features for {R}: Standardized Support for Spatial Vector Data},
-  journal = {The R Journal},
-  year    = {2018},
-  volume  = {10},
-  number  = {1},
-  pages   = {439--446},
-  url     = {https://journal.r-project.org/archive/2018/RJ-2018-009/index.html},
 }
 
 @Book{Press.etal2007,
@@ -108,14 +97,6 @@
   address   = {Cambridge},
   edition   = {3rd},
   isbn      = {0-521-88068-8},
-}
-
-@Manual{R:mvtnorm,
-  title  = {\texttt{mvtnorm}: Multivariate Normal and t Distributions},
-  author = {Alan Genz and Frank Bretz and Tetsuhisa Miwa and Xuefei Mi and Torsten Hothorn},
-  year   = {2018},
-  note   = {R package version 1.0-8},
-  url    = {https://CRAN.R-project.org/package=mvtnorm},
 }
 
 @Manual{R:base,
