### 2019-01-31 0d2e057
ensure that polyCub functions are found
