+++ R/plotpolyf.R
@@ -16,11 +16,13 @@
 ##' or \code{\link{image}}.
 ##'
 ##' @param polyregion a polygonal domain.
-##' The following classes are supported: \code{"\link[spatstat]{owin}"},
-##' \code{"\link[rgeos:gpc.poly-class]{gpc.poly}"},
+##' The following classes are supported:
+##' \code{"\link[spatstat]{owin}"} from package \pkg{spatstat},
+##' \code{"\link[rgeos:gpc.poly-class]{gpc.poly}"} from \pkg{rgeos} (or
+##' \pkg{gpclib}), as well as
 ##' \code{"\linkS4class{SpatialPolygons}"}, \code{"\linkS4class{Polygons}"},
-##' and \code{"\linkS4class{Polygon}"}
-##' (for these we have an internal \code{\link{xylist}} method).
+##' and \code{"\linkS4class{Polygon}"} from package \pkg{sp}.
+##' (For these classes, \pkg{polyCub} knows how to get an \code{\link{xylist}}.)
 ##' @param f a two-dimensional real-valued function.
 ##' As its first argument it must take a coordinate matrix, i.e., a
 ##' numeric matrix with two columns, and it must return a numeric vector of
