### 2019-01-31 695cfd3
use method-specific examples
