### 2019-01-31 0323d96
use upcoming JOSS paper for citation("polyCub")
