### 2019-01-31 #1056 @bastistician
I have
* implemented the suggested improvements and increased the package version to 0.7.1
* not yet tagged that release nor submitted to CRAN
* made minor modifications in the paper (referencing the new vignette), see the updated article proof

@wrathematics : May I include your name as a reviewer in the package [DESCRIPTION](https://github.com/bastistician/polyCub/blob/0323d96a0f7913571463995d2686f0b74341c6d9/DESCRIPTION) (with role="rev")?

@karthik : How do we proceed? Is there an intended release schedule at JOSS? I would like the package to be cited using this JOSS publication, so I have included a reference in the package [CITATION](https://github.com/bastistician/polyCub/blob/0323d96a0f7913571463995d2686f0b74341c6d9/inst/CITATION) file. However, the DOI will only be registered later, and the volume and issue number are probably not yet fixed (or can I assume "4(33)" as indicated in the footnote of the current proof?). Should I release the package on CRAN prior to or after JOSS acceptance and archiving on Zenodo?
