### 2019-02-04 #1056 @karthik
@bastistician Can you please deposit the latest version of your package on Zenodo and share the DOI with me here? 🙏
