### 2019-02-07 #1056 @bastistician
@karthik Today, I have released the package on [CRAN](https://CRAN.R-project.org/package=polyCub) (version 0.7.1) and created a corresponding [GitHub release](https://github.com/bastistician/polyCub/releases/tag/v0.7.1), which has subsequently been pushed to Zenodo. Here is the DOI:

https://doi.org/10.5281/zenodo.2559486
