### 2019-02-12 #1056 @arfon
> LGTM. EICs, this is ready to accept.

Thanks @karthik. In the future can you ping @openjournals/joss-eics?
